Bump octavia-tempest-plugin to 1.7.0 till victoria
https://review.rdoproject.org/r/c/rdo-jobs/+/32279 moves the
sc10 tests to smoke tests which exists in octavia-tempest-plugin
1.5.0 onwards.

For train, we have to support py2 and py3 both so sticking with
1.6.0 version.

Octavia-tempest-plugin exists with 1.3 in train and 1.4 in ussuri
leading to failing sc10 jobs with no test regex found.

Bumping to latest version will fix the issue and improves the
tests for all releases.

Change-Id: Ib37f7645b060d39179223faaf7f7aedde33ec43e
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>