WIP: Bump tempest to 27.0.0 in Ussuri
The new release of tempest contains fixes and improvements
good having in the latest stable release.
The version of tempest is bumped to the following commit [1].

The reason we pin tempest version to that commit is the commit [1]
contains a fix for a bug [2] introduced in the latest commits [3]
of 28.0.0 tag.

[1] https://opendev.org/openstack/tempest/commit/10a649b098279220468532dfc92db3779fed492a
[2] https://bugs.launchpad.net/tempest/+bug/1939049
[3] https://review.opendev.org/c/openstack/tempest/+/797466/

Depends-On: https://review.rdoproject.org/r/#/c/27710/
Depends-On: https://review.rdoproject.org/r/c/openstack/tempest-distgit/+/35079
Change-Id: Id7b590ebd66a3e55bfd24b6427a2a5b5484366ec