Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> Dup to https://review.rdoproject.org/r/c/rdoinfo/+/32936
> 
> IIRC the idea is to:
> 1. Drop periodic-tripleo-ci-centos-7-scenario010-standalone-train (see https://review.rdoproject.org/r/c/rdo-jobs/+/32967)
> 2. Bump octavia-tempest-plugin to 1.7.0 in all supported versions (>= train, where train is now py3 only)

I have separated this review , we are handling train case separately, As for train, we need to build octavia-tempest-plugin-1.7.0 both for c7 and c8 as other packages are also built for c7 and c8 on DLRN server for train and we have integration promotion line running for c7 and c8 for train. So in order to unblock that I have separated this review.

Patch-set: 2