Bump octavia-tempest-plugin to good known hash for train
https://review.rdoproject.org/r/c/rdo-jobs/+/32279 moves the
sc10 tests to smoke tests which exists in octavia-tempest-plugin
1.5.0 onwards.

For train, we are dropping c7 sc10 octavia job from the pipeline.

Octavia-tempest-plugin exists with 1.3 in train and 1.4 in ussuri
leading to failing sc10 jobs with no test regex found.

Bumping to good known hash before
https://review.opendev.org/c/openstack/octavia-tempest-plugin/+/747959
(which adds httpx as a deps) will fix the issue and improves the
testing.

We also need to make sure octavia-tempest-plugin with this hash
works for both py2 and py3.

Depends-On: https://review.rdoproject.org/r/c/rdo-jobs/+/27710

Change-Id: Ib37f7645b060d39179223faaf7f7aedde33ec43e
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>