[WIP] Bump octavia-tempest-plugin to 1.7.0 for train
https://review.rdoproject.org/r/c/rdo-jobs/+/32279 moves the
sc10 tests to smoke tests which exists in octavia-tempest-plugin
1.5.0 onwards.

For train, we are dropping sc10 octavia job from the pipeline.

Octavia-tempest-plugin exists with 1.3 in train and 1.4 in ussuri
leading to failing sc10 jobs with no test regex found.

Bumping to latest version will fix the issue and improves the
tests for all releases.

We also need to make sure octavia-tempest-plugin 1.7.0 also
works on py2 c7 as for train DLRN builds for c7 and c8 builder.

Change-Id: I77c0b14216f7abe988b47d7ae510e61ae62f187e
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>