Update patch set 1
Patch Set 1:

This, as I understand it, is an optional requirement for Anaconda based deployment. *We* as downstream teams do not have the bandwidth available to go add another thing to our consciousness nor further subdivide our attention and resources. I truly don't understand what the ask in this thread is, but if we can't consume what is published without taking further steps, then we simply should not merge this patch and move on in life. Especially since the requirement change is for Xena, not Wallaby, and the feature to use the code, as far as I'm aware has not merged yet, nor would it ever be backported to Wallaby.

Patch-set: 1
Reviewer: Gerrit User 511 <511@270e2033-b340-4cff-9539-693957ebf0e7>