Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> (1 comment)
> 
> > Patch Set 1: Code-Review-1
> > 
> > I see 2 issues with unpinning tempestconf:
> > * py2/py3 compatibility, master code is not py2 compatible anymore
> > * we have gates only until back to train (and in gates it's executed still by py3, however for packaging purposes the project needs to f.e. pass unit tests executed by py2 - we had to do some workarounds in order to package tempestconf for train last time https://review.rdoproject.org/r/c/openstack/tempestconf-distgit/+/28823)
> 
> We pin projec

Thank you both for your help and for clarifying the issue to me. Martin created a new tag that we could use and he tested pinning the new tag on stable/train: https://review.rdoproject.org/r/c/rdoinfo/+/33233

This should be a good approach to get some changes we need in stable/train! :) 

I'll abandon this patch and follow up on Martin's

Patch-set: 1