Check if projects sorted in buildsys-tags files
We add a unit test which checks if projects in buildsys-tags
files are well sorted.
As the script 'add-buildsys-tags.py' check the projects order when
adding/editing a project and fix it if not, we need to ensure that
this order is correct.
This should avoid us to have case like [1] with novnc project change.

[1] https://review.rdoproject.org/r/c/rdoinfo/+/32318/1/buildsys-tags/cloud8s-openstack-wallaby-candidate.yml

Change-Id: I204e35685a18743f5d38cdd2faedde5b874144ce