Set main as source-branch for some puppetlabs repos
[1] set the source-branch for some puppetlabs repositories to "main",
following their upstream change. However, it was only done for the
xena-uc tag.

This commit extends that to the xena tag, and also sets it at the
project definition level, to signal that there is no "master" branch
at all. This will also help DLRN jobs when no tag is defined.

[1] - https://review.rdoproject.org/r/c/rdoinfo/+/33695

Change-Id: Ie4f9c7559ca213557d8b03c6f76414d234cccf1a