Pin Barbican to working commit
With [1] barbican needs sqlalchemy >= 1.4.0 but currently
we have 1.3.23 in xena repos.
To match u-c and to make barbican work attempted to
update sqlalchemy to 1.4.18[2], but it fails as promoted
barbican commits do not work with sqlalchemy-1.4.18, also
gnocchi and watcher faced some issues due to it.

Barbican needs to be get fixed in a backward compatible
way to get sqlalchemy updated. Until than proposing
pinning of barbican to working commit.

[1] https://review.opendev.org/c/openstack/barbican/+/796284
[2] https://review.rdoproject.org/r/c/rdoinfo/+/34221

Related-Bug: lp#1932527
Change-Id: I7c3c15f3287b58d75298bdb5b67f3b49ff8c8150