Pin python-networking-l2gw-tests to previous commit
This pin commit is 6 commit older than the one we want to
replace [1]. The difference is not relevant as the project is
retired.
We can't build the post commit 82e3d07ea410e1489c6a83e5a4b9557d52643d41
as the project has been archived in Github before the
commit was pushed. So, there is no tarball built for this commit in
Github [2].
Note that in buildsys-tags, Wallaby is using
python-networking-l2gw-tests-tempest-0.1.0-2.2e265b5git.el8

[1] https://opendev.org/x/networking-l2gw-tempest-plugin/commits/branch/master#
[2] https://github.com/openstack/networking-l2gw-tempest-plugin/archive/82e3d07ea410e1489c6a83e5a4b9557d52643d41.tar.gz#/python-networking-l2gw-tests-tempest-82e3d07.tar.gz

Change-Id: Ia814f3c05d41518f8dfeeb778f1b80209e4e767e