Set repositories for dependencies according to the new process
With the new dependencies workflows, we are maintaining the deps
distgits in review.rdoproject.org repos deps/<package_name>.

This patch is setting the repos for the distgits:

- distgit and master-distgit: distgits repo in review.r.o.
- upstream: distgit repo in fedora (usually we take the content from
  fedora to rebuild in centos.
- centos-distgit: distgits in centos.

This new centos-distgit is a new attribute for informative purpose. We
may implement a new feature in `rdopkg clone` where we can add
additional repos to do something like `rdopkg clone python-eventlet` and
we get the distgits in fedora, rdo and centos cloned.

Note this patch is still maintaining both rdo-dependency and
fedora-dependency. I think, once we finish the implementation of the new
process, we'll be able to merge in a single one, but we need to fix some
jobs before which still behave differently in fedora or rdo deps.

Change-Id: I06e1fc2df5576f73185db251a3b9b458249c469b