Update patch set 2
Patch Set 2:

Initial bootstraping of any of the repos is hard. Ideally this one would live in deps as it's not version specific and we can provide a basic way to install.  tripleo-operator-ansible will likely get a rhos-bootstrap to handle this as well.  What I don't want is this ending up in a version specific thing under tripleo. So would common be better or should I just leave my updated to the tripleo component

Patch-set: 2