Pin scciclient in ussuri to last known good commit
scciclient has stable branches in victoria and wallaby but not in
ussuri, so we are ahead in ussuri that we are in wallaby or victoria
which is abnormal. Follow master branches in stable releases is not
recommended unless it's requested by maintainers, so let's pin to last
known good branch (it's FTBFS now).

Change-Id: I98442eb4b5ee8a13b4849fb2e3f3308ad0966564