Pin tempest for Wallaby and Xena
Currently multiple tempest tests are failing due to changes
in ssh key type.

Tempest had got the support for
the same but fixes are not yet available in wallaby as it
is pinned to stable tags.

openssl new version in CentOS 9-stream dropped support for
'rsa' keys, So we need to use the supported key.

For that, we need to pin tempest at proper commit containing
those fixes till we get a new release.

* https://review.opendev.org/c/openstack/tempest/+/822560

The above issue is tracked here: https://bugs.launchpad.net/tripleo/+bug/1963907

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: I995851155fc2ba8f1c5a86746a5fae319b5b268a