Pin neutron-tempest-plugin for wallaby and Xena
Recent release of openssl broke the neutron tempest tests on
wallaby CS9 due to key change.

In order to fix that, we bumped the tempest to a particular commit[1].
We also need to pin neutron-tempest-plugin at good hash.

https://review.opendev.org/c/openstack/neutron-tempest-plugin/+/832640
and https://review.opendev.org/c/openstack/neutron-tempest-plugin/+/832911
fixes the issue and also makes it compatible with
wallaby release.

[1]. https://review.rdoproject.org/r/c/rdoinfo/+/40228

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: I9b2f2a14e5b9ee8e85da58b37d84c350eb5af126