Add buildsys-tags for Zed
This is part of the Yoga branching. At this poing we can diverge
dependencies between zed and yoga, so in this review I'm adding Zed
release to rdoinfo and adding the dependencies builds we have now in
yoga to zed.

Next step, will be to also switch tags from yoga to zed, but i'll do
that in a separate review.

Change-Id: Icd3a73916e35352462797b72d0e88896ab341bf5