Add tags for Zed
This is part of the Yoga branching. At this poing we can diverge
dependencies between zed and yoga, so in this review I'm switching
master tags to zed ones and moving yoga to pinned releases.

Note that we need to coordinate this change with the one changing
RDO-Trunk to use zed-uc for master-uc.

Change-Id: I4b83e06cefb1434671cc1957f2e9e74e7d5b4ae1