Update jinja2 to 3.0.1 in Zed
Some projects are now requiring jijna2>=3.0.0. Note that this is
overriding the jinja2 package included in CS9 (2.11.3) and has some
non-backwards compatible changes.

Note that the build we are promoting is provided already in Storage SIG
repos for Ceph, and we must be compatible with them (actually, we are
already pulling it in CI from storage repo).

Change-Id: Ica952d4b97189006eaaecef0ecc54517dda891d4