Pin neutron-vpnaas to train-em commit
neutron-vpnaas has been EOL'd in train and ussuri and branches have been
removed so dlrn is falling back to master and trying to follow it. This
patch is pinning neutron-vpnaas to the commit which was starting point
of stable/train branch (the latest we can set in train after the branch
removal). Note that we have older builds in DLRN so this effectively
means it will not try to build it again.

Change-Id: I340c7cd6cb57ccbf7271f140e3f086950730242d