Pin neutron-vpnaas in ussuri and cinderlib in ussuri and train
neutron-vpnaas and cinderlib have been EOL'd in train and ussuri and the
stable branches have been removed upstream so dlrn is falling back to
master and trying to follow it (and failing).

This patch is follow-up of [1] and pinning both projects in train and
ussuri to the commits that were used to cut stable branches from master.
Note that we have newer builds in RDO Trunk train and ussuri so DLRN
will not try to build anything after this patch.

[1] https://review.rdoproject.org/r/c/rdoinfo/+/46076

Change-Id: Ia1d245fe789d89b2e1c0528070da9ba2f8bf7144