Add python-uhashring and python-binary-memcached
binary-memcached was added as dependency to oslo-cache during zed cycle
but we missed to add it so this is breaking the memcached driver
feature. This patch is addiong both python-binary-memcached and
python-uhashring which is a depencency for it.

Note we are currently in the process of adding them to Fedora too.

Change-Id: Ief4125f25045945b84fce51477b07720a4c5a7fe