Revert ironic-tests-tempest to 2.2.0 in Train
Version 2.3.0 includes [1] which moves some actions to use system scoped
tokens. This had a problem with tempest which was fixed in version 28.0.0 [2]
while we have 26.1.0 in Train if create_isolated_networks option is
enabled. This is breaking p-o-i jobs while tripleo jobs have been
working fine as create_isolated_network is disabled.

[1] https://github.com/openstack/ironic-tempest-plugin/commit/3a07c4d5ae3cb277e860d710bee0adb1c2555ee6
[2] https://github.com/openstack/tempest/commit/5dcdd41ddafa2e455ab0dd0ce0b1d3f3e32caa04

Change-Id: I9362fd085855447b208eed6e57d383a6ce3472f4