Bump designate-tmepest-plugin to 0.15.0 in Z, Y, X and W
The 0.15.0 version is not the marked as wallaby-last
(that would be 0.16.0), but it seems that 0.16.0 contains
a change which broke the tests:
https://review.opendev.org/c/openstack/designate-tempest-plugin/+/860116

and the relevant fix was merged only after the 0.16.0 release:
https://review.opendev.org/c/openstack/designate-tempest-plugin/+/862804

which is also the last available release right now,
which means the tag can't be bumped further in the still
supported releases, from xena onward.

There are no many other changes between 0.15.0 and 0.16.0,
so 0.15.0 should still be a good improvement from the current tags.

Change-Id: I912aa2a1c1c42810bb7e3c373c2daf9e20cf3713