Add ansible-core as depencency
The ansible-core package has been rebuilt on python 3.11 in CentOS Stream [1].
That is making tripleo to fail at deploying time [2]. Trying to fix it
with some patches in tripleoclient and tripleo-ansible have shown that
having different python versions on ansible control side and
openstack/tripleo packages is not working as some code running in
control mode (on py3.11) requires python modules which are bult on
python 3.9.

It has been decided to rebuild ansible-core from CentOS Stream build on
python 3.9 in RDO and override the package in CentOS Stream repos by
bumping epoch or some alternative technique and shipping it as a
dependency. Note that we will not rebuild it from Fedora but from the
spec in CentOS Stream.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=2169466
[2] https://bugzilla.redhat.com/show_bug.cgi?id=2170553

Change-Id: I44cf6c27bb14b94c7f663f07be98c5ecef92c179