Add new attribute to the releases with upstream identifier
Until Zed release, upstream stable branches name was `stable/<release>`.
Since Antelope, releases have a name and a identification code based on
the year [1]. While the name is used to identify the release
deliverables folder in the openstack/releases project, the year code is
used in branch names `stable/<identifier>` so we need to know it. This
patch is adding a new attribute `identifier` for each release that will
contain the year-based identifier for each release.

[1] https://governance.openstack.org/tc/reference/release-naming.html

Change-Id: I7df0e4a80bef8981e50c02709d118c0b0acca8c7