Switch RDO master to bobcat release (part 2)
Follow-up of https://review.rdoproject.org/r/c/rdoinfo/+/47877

This patch is moving master tags to bobcat and bobcat-uc (for
centos9-master and centos9-master-head) and making antelope tags to
follow upper-constraints (similar to latest state of antelope-uc).

Note that we need to coordinate the merge of this change with a change
in rdo trunk config to use the new tags.

Change-Id: Ib17454ade24a23713a8f5ea0cc1a1556ce9bb4bd