Pin puppet-stdlib to last promoted commit
Upstream maintainers of puppetlabs-stdlib [1] are
currently removing bunch of deprecated bits leading
to failures in promotion pipeline. As a result, last
promotion was 10 days ago, and it may take time to
fix all the issues which appear one by one.
So, we pin it to last promoted commit for now, in order
to have a promotion. In the meantime, we'll fix all the
issues upstream with latest stdlib base code, and then
we'll unpin it.

[1] https://github.com/puppetlabs/puppetlabs-stdlib

Change-Id: Ib932209bf06f71802af0c90df386da6f859e9291