Tripleo removal from RDO Master (2 - remove and reorganize)
TripleO jobs have been removed upstream [1] and TripleO is deprecated
[2].

This patch is removing TripleO packages in Bobcat release. Note this
change will not affect to RDO Zed which will keep tripleo packages.

I am applying the same changes as for Antelope in [3] to remove tripleo
packages and move the ones that we want to keep to other components.

[1] https://lists.openstack.org/pipermail/openstack-discuss/2023-March/033080.html
[2] https://review.opendev.org/c/openstack/governance/+/877132
[3] https://review.rdoproject.org/r/c/rdoinfo/+/47883

Change-Id: Iae52ed35f24afff97decc398f002fc1145d2ffc5