Tripleo removal from RDO Master (uc builder)
This patch is removing TripleO packages in Bobcat release. Note this
change will not affect to RDO Zed which will keep tripleo packages. The
removal for the non-uc builder will be sent in a separated review.

TripleO jobs have been removed upstream [1] and TripleO is deprecated
[2].

I am applying the same changes as for Antelope in [3].

[1] https://lists.openstack.org/pipermail/openstack-discuss/2023-March/033080.html
[2] https://review.opendev.org/c/openstack/governance/+/877132
[3] https://review.rdoproject.org/r/c/rdoinfo/+/47883

Change-Id: Iae52ed35f24afff97decc398f002fc1145d2ffc5