Tripleo removal from RDO Master (1 - create puppet component)
TripleO jobs have been removed upstream [1] and TripleO is deprecated
[2].

In order to remove the tripleo component following the same approach as
for Antelope [3], we need to first get the puppet component created in
the RDO Trunk builders centos9-master-uc and centos9-master. This patch
is moving the first package to that component. Note that moving all
packages together in this review makes the CI job to fail, so the rest
of packages will be moved in the next review.

[1] https://lists.openstack.org/pipermail/openstack-discuss/2023-March/033080.html
[2] https://review.opendev.org/c/openstack/governance/+/877132
[3] https://review.rdoproject.org/r/c/rdoinfo/+/47883

Change-Id: I47a877646a75436ed3888587a4894ae2a727fb40