Revert "Revert "Bump eventlet to 0.33.3 in Antelope""
This reverts commit 762f27fa1c3b82302637f59622bf18000a4d315b.

We first bump eventlet to 0.33.3 as the new commit of ovn-bgp-agent
was failing to build [1], and the same commit (ovn-bgp-agent has
master branch only) was built sucessfully in master which has
eventlet 0.33.3. But then, I noticed eventlet is currently pinned to
0.33.1 in Antelope u-c [2], so we reverted the commit [3].

After digging into it, the real issue of the failure of ovn-bgp-agent
build is because of the bump of python3-dns [4] from BaseOS on
2023-06-27. dnspython 2.3.0 is not compatible with eventlet<0.33.3 [5][6].

Plus, dnspython is pinned to 2.2.1 [7] in u-c.
So we are forced by CentOS official repo to be above u-c version of
dnspython, and this implies bumping eventlet to 0.33.3 which is above
u-c version as well.

[1] https://review.rdoproject.org/r/c/openstack/ovn-bgp-agent-distgit/+/49141
[2] https://github.com/openstack/requirements/blob/d450d25d953cc53e902f83c6a0123efa2536faeb/upper-constraints.txt#L158
[3] https://review.rdoproject.org/r/c/rdoinfo/+/48957
[4] https://mirror.stream.centos.org/9-stream/BaseOS/x86_64/os/Packages/python3-dns-2.3.0-2.el9.noarch.rpm
[5] https://github.com/rthalley/dnspython/issues/885
[6] https://github.com/eventlet/eventlet/issues/781
|7] https://github.com/openstack/requirements/blob/d450d25d953cc53e902f83c6a0123efa2536faeb/upper-constraints.txt#L549

Change-Id: Ifa3f68757f48739d9e735bcefac44095efad2b25