Pin scciclient to last known-good hash in master
The latest commit [1], tagged as 0.15.0 migrate pynsnmp to
pysnmp-lextudio because of an incompatibility between pysnmp and
latest pyasn1.
We consume python3-pyasn1-0.4.8-6.el9.noarch from Appstream repo
which is still compatible with our current python3-pysnmp.
We pin to 0.14.0 for now, this should give us time to package
pysnmp-lextudio. Note that the latest known-good hash is tagged
0.14.0, one commit before.

[1] https://review.opendev.org/c/x/python-scciclient/+/885942

Change-Id: I1ebeddf984b84a81db48051f0148b35b44d5968f