Bump openstackclient post u-c 4.0.2 in Train
In order to include the fixes of rhbz#2027879 [1], we have
to bump to post 4.0.2 (u-c version currently [2]). Even though
we keep following libs/clients from upper-constraints, I think
we can safely bump up to this commit.
Note that this will apply only for Trunk repo (we don't build
CloudSIG bits for Train anymore).

[1] https://bugzilla.redhat.com/show_bug.cgi?id=2027879
[2] https://github.com/openstack/requirements/blob/c5e371c0153b5d22200c810d1c84e88d6df9f072/upper-constraints.txt#L555

Change-Id: Ia5cfdc7578875592cd3d44519ca2046d03cf1acb