Switch RDO master to caracal release (part 1)
This patch is moving master tags to caracal and caracal-uc (for
centos9-master and centos9-master-head) and making bobcat tags to
follow upper-constraints (similar to latest state of bobcat-uc).

Note that we need to coordinate the merge of this change with a change
in rdo trunk config to use the new tags.

Change-Id: I2b0a00c061deab5abbf7b39e29ebe900d8f3a648