Pin manila tempest plugin in Antelope
Since the bump of all tempest plugins in Antelope with
[1], the puppet-promotion pipeline is blocked.
After investigation, manila tempest plugin 2.0.0 is the
blocker. The commit [2] needs some patches [3] to be
backported in stable/antelope. However, the backport operation
is too big with bunch of conflicts to be resolved.
In the meantime, I think we should pin the plugin to the commit
before the one failing.

[1] https://review.rdoproject.org/r/c/rdoinfo/+/51422
[2] https://review.opendev.org/c/openstack/manila-tempest-plugin/+/888618
[3] https://review.opendev.org/q/topic:%22bp/allow-locking-shares-against-deletion%22+project:openstack/manila

Resolves: https://issues.redhat.com/projects/RDO/issues/RDO-206
Change-Id: Ie57ba81067a6dc16a48203e85982942c04c31520