Add upstream_branch data for OpenStack releases
After the implementation of the unmaintained branches policy [1]
branch names change in stable releases once they transition to
unmaintained status.

In the past, we assumed that branch name was stable/<release> for
stable releases until they are EOL'd [2] but that is not longer
correct.

This patch is adding the informatin about the default branch name for
all openstack releases as part of the releases metadata so that we can
use it in the automation via rdopkg or distroinfo.

Issue: https://issues.redhat.com/browse/RDO-257

[1] https://docs.openstack.org/project-team-guide/stable-branches.html
[2] https://github.com/softwarefactory-project/rdopkg/blob/e5721d98c878af38179c0c265a0706e3872dc0b9/rdopkg/actionmods/rdoinfo.py#L54

Change-Id: I087f9c5a9fc907286d7def30e4379fc2114388af