Split Designate out from the Network component
In order to ensure that Designate issues do not cross with Neutron
issues (and vice versa) split Designate out to a different component.
This is only for Antelope onwards, since we do not care about Zed and
before for the purposes of this activity.  This split was done following
the prior art contained in [1].

[1] https://review.rdoproject.org/r/c/rdoinfo/+/51426

Change-Id: I5a0319cfde043c4af87a4ffcc748c070c5b0a09e