Switch RDO master to Dalmatian release (2/2)
Follow-up of https://review.rdoproject.org/r/c/rdoinfo/+/52473

This patch is moving master tags to dalmatian and dalmatian-uc (for
centos9-master and centos9-master-head) and making Caracal tags to
follow upper-constraints (similar to latest state of caracal-uc).

Note that we need to coordinate the merge of this change with a change
in rdo trunk config to use the new tags.

Change-Id: I1be64257f71c03b5f2386e104c8436ace7044106