Switch RDO master to Epoxy release (2/2)
Follow-up of https://review.rdoproject.org/r/c/rdoinfo/+/54764

This patch is moving master tags to epoxy and epoxy-uc (for
centos9-master and centos9-master-head) and making Dalmatian tags to
follow upper-constraints (similar to latest state of dalmatian-uc).
Note that we need to coordinate the merge of this change with a change
in rdo trunk config to use the new tags.

Change-Id: Ifa0f27abd8607ae3f030ccead593833cc5f42b91