Pin sushy-tools to last well-known hash commit in Bobcat
The project bumped pbr to 6.0.0 [1] but it's still 5.11.1 in
u-c for this release [2]. We cannot uncap pbr in %prep section
as the failure occurs when DLRN guesses the version to build the
SRPM, so before the build itself.
I think it's safe to pin to this commit as this release is old
enough and diverge too much from master.

[1] https://review.opendev.org/c/openstack/sushy-tools/+/933594
[2] https://github.com/openstack/requirements/blob/f18d00dc711735bdd98d985513e4a3edec44dd04/upper-constraints.txt#L44

Change-Id: I2d1d98c88bcfacecadb8a59ad92110d2e105df25