Pin sushy-tools to last well-known hash commit in Antelope
The project bumped pbr to 6.0.0 [1] but it's still 5.11.1 in
u-c for this release [2]. We cannot uncap pbr in %prep section
because the failure occurs when DLRN guesses the version to build
the SRPM, so before the build itself.
I think it's safe to pin to this commit as this release is old
enough and diverge too much from master.

[1] https://review.opendev.org/c/openstack/sushy-tools/+/933594
[2] https://github.com/openstack/requirements/blob/d960c51d45b79d29c7b231be2fe7bc648fb226e6/upper-constraints.txt#L45

Change-Id: I4e7104fa0a11ca3c752d645ca3400dcdc463e60d