Pin sushy-tools to last well-known hash commit in Zed
The project bumped pbr to 6.0.0 [1] but it's still 5.10.0 in
u-c for this release [2]. We cannot uncap pbr in %prep section
because the failure occurs when DLRN guesses the version to build
the SRPM, so before the build itself.
I think it's safe to pin to this commit as this release is old
enough and diverge too much from master.

[1] https://review.opendev.org/c/openstack/sushy-tools/+/933594
[2] https://github.com/openstack/requirements/blob/1a3b0b7508e206c1863f5a700c1e359728010a61/upper-constraints.txt#L45

Change-Id: Iff72ee76d133b59d7b8dce0798d7e51472e21b10