Create new component optimize for watcher service
In order to isolate watcher service and avoid issues on it to block
other packages in common component for systems using per-component CI,
this review is creating a new component optimize (watcher is intended
for infra optimization) and moving watcher service into it.

This review is doing the change globally for all releases from antelope
up to master. A dummy commit will be sent to watcher and a package in
common after is merged to force both components to be updated.

Change-Id: I6a5905e97cb8a6cf2009ab6df69a0ec1e67687a5