Switch RDO master to Flamingo release (2/2)
Follow-up of https://review.rdoproject.org/r/c/rdoinfo/+/57262

This patch is moving master tags to flamingo and epoxy-uc (for
centos9-master and centos9-master-head) and making Epoxy tags to
follow upper-constraints (similar to latest state of epoxy-uc).
Note that we need to coordinate the merge of this change with a change
in rdo trunk config to use the new tags.

Change-Id: I9f62081b98fa7891876c8524180e0661d75fa972